inventory-control-ml/backend/requirements.txt
Django==4.0.6
djangorestframework==3.13.1
pandas==1.4.2
numpy==1.22.3
scikit-learn==1.0.2
auto-sklearn==0.14.7
tpot==0.11.7
prophet==1.1
shap==0.40.0
joblib==1.1.0
python-dateutil==2.8.2
scikit-learn==1.0.2
tpot==0.11.7